will run fnc2fever using stand alone lex. first/now on toy data

+++ mithun_scripts/get_fever_fnc_data.sh
@@ -4,17 +4,15 @@
 
 
 
-
-
-
 #######fevercrossdomain mod1 (training and dev will be in fever (with 4 labels), and test on fnc-dev partition)
-if [ "$TASK_TYPE" = "mod1" ] && [ "$TASK_NAME" = "fevercrossdomain" ] && [ "$SUB_TASK_TYPE" = "figerspecific" ]; then
+if [ "$TASK_TYPE" = "lex" ] && [ "$TASK_NAME" = "fevercrossdomain" ] && [ "$SUB_TASK_TYPE" = "figerspecific" ]; then
 
-echo "found task type is mod1 and task name as fever cross domain"
 
 echo $DATA_DIR
 mkdir -p $DATA_DIR
+echo "found task type is lex and task name as fever cross domain"
 
+exit
 
 
 FILE=$DATA_DIR/train.tsv
@@ -324,6 +322,57 @@ else
       #wget https://osf.io/4n7b6/download -O $FILE
 
       lexicalized version of fever indomain dev partition. use this when you are training on lex
+      wget https://osf.io/xdbh6/download -O $FILE
+
+
+
+
+fi
+
+fi
+
+
+
+
+####################################for cross domain student teacher, there will be two training files.-one for lex and another for delex
+#fnccrossdomain means, train on fnc and test on fever
+# note that when training on fnc and testing on fever, the total numbere of labels will be that in fever. So in this case, when we
+# use fnccrossdomain, data, it is a data where 4 labels (agree, disagree, discuss, unrelated) has been reduced to 3 for fever (agree, disagree,nei)
+
+
+if [ "$TASK_TYPE" = "lex" ] && [ "$TASK_NAME" = "fnccrossdomain" ] && [ "$SUB_TASK_TYPE" = "figerspecific" ]; then
+    echo "found task type to be lex, taskname to be fnccrossdomain and subtasktype to be figerspecific"
+
+echo $DATA_DIR
+mkdir -p $DATA_DIR
+
+#train1.tsv will be the lexicalized  fnc train data but with 3 labels (to be same as that of fever).i.e 40904 datapoints. 3 labels, agree, disagree, nei
+FILE="$DATA_DIR/train.tsv"
+if test -f "$FILE";then
+    echo "$FILE exists"
+else
+    wget https://osf.io/dwef7/download -O $FILE
+fi
+
+
+FILE="$DATA_DIR/dev.tsv"
+if test -f "$FILE";then
+    echo "$FILE exists"
+else
+    #dev.tsv will be the lexicalized  fnc dev data but with 3 labels
+    wget https://osf.io/a2m6e/download -O $FILE
+
+fi
+
+
+#note that we are  replacing the test partition with cross domain dev partition
+
+FILE="$DATA_DIR/test.tsv"
+if test -f "$FILE";then
+echo "$FILE exists"
+else
+
+      #lexicalized version of fever indomain dev partition. use this when you are training on lex
       wget https://osf.io/xdbh6/download -O $FILE
